requirements.txt
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk
opentelemetry-api
